allow use of parted instead of sfdisk
This change allows using parted to perform the partitioning of
the block device instead of sfdisk. It defaults to continuing
to use sfdisk and it's up to the user to decide if they wish
to use parted.

This was added to work around an issue sfdisk was having when
partitioning a RADOS block device, the ioctl to the device
asking the kernel to re-read the partition table was failing.
This resulted in sfdisk exiting with error code 1, even though
the partitioning was successful.