cloud-init hook
This utilizes and requires cloud-init to be installed and enabled on an example.
This pulls several bash variables to generate a NoCloud cloud-init config. On
boot, cloud-init will read the config and set the hostname, user/group, ssh key
and dns settings for the instance. In addition it will grow the partition and
rootfs if those tools are installed and setup properly.

This hook is still a work in progress, but this is the first iteration of it.